layered: match Java float precision in OER xpos and LGraphUtil resize
Java's OrthogonalEdgeRouter uses float (f32) for xpos, startPos, and
routingWidth. Rust was using f64, causing sub-pixel coordinate drift.
Truncate through f32 at each accumulation step to match Java's behavior.

Also match f32 precision in LGraphUtil::resize_node for width/height
ratios and diffs (Java lines 90-93).

740 → 746 matches (+6).

+++ plugins/org.eclipse.elk.alg.layered/src/org/eclipse/elk/alg/layered/graph/l_graph_util.rs
@@ -48,10 +48,11 @@ impl LGraphUtil {
             Err(_) => return,
         };
 
-        let width_ratio = new_size.x / old_size.x;
-        let height_ratio = new_size.y / old_size.y;
-        let width_diff = new_size.x - old_size.x;
-        let height_diff = new_size.y - old_size.y;
+        // Java uses float (f32) for these variables. Truncate to match.
+        let width_ratio = (new_size.x / old_size.x) as f32 as f64;
+        let height_ratio = (new_size.y / old_size.y) as f32 as f64;
+        let width_diff = (new_size.x - old_size.x) as f32 as f64;
+        let height_diff = (new_size.y - old_size.y) as f32 as f64;
 
         if move_ports {
             for port in ports {
